Core/SSH: ssize_t/size_t to int casts are made explicit to make compilers happy.
git-svn-id: svn://ultimatepp.org/upp/trunk@15443 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/SSH/SFtp.cpp
@@ -135,7 +135,10 @@ int SFtp::Read(SFtpHandle handle, void* ptr, int size)
 {
 	int sz = min(size - done, ssh->chunk_size);
 
-	int rc = libssh2_sftp_read(handle, (char*) ptr + done, sz);
+	int rc = static_cast<int>(
+		libssh2_sftp_read(handle, (char*) ptr + done, size_t(sz))
+		);
+
 	if(!WouldBlock(rc) && rc < 0)
 		SetError(rc);
 	if(rc > 0) {
@@ -152,7 +155,10 @@ int SFtp::Write(SFtpHandle handle, const void* ptr, int size)
 {
 	int sz = min(size - done, ssh->chunk_size);
 
-	int rc = libssh2_sftp_write(handle, (const char*) ptr + done, sz);
+	int rc = static_cast<int>(
+		libssh2_sftp_write(handle, (const char*) ptr + done, size_t(sz))
+		);
+
 	if(!WouldBlock(rc) && rc < 0)
 		SetError(rc);
 	if(rc > 0) {
